pyproject.toml
[project]
name = "disaster-tweets-nlp"
version = "0.1.0"
description = "PyTorch Lightning + Hydra + DVC + MLflow project for Disaster Tweets classification"
readme = "README.md"
requires-python = ">=3.11,<4.0"
authors = [{ name = "Your Name", email = "you@example.com" }]

dependencies = [
  "pandas>=2.2.2,<3.0.0",
  "numpy>=2.0.1,<3.0.0",
  "scikit-learn>=1.5.1,<2.0.0",
  "torch>=2.4.0,<3.0.0",
  "pytorch-lightning>=2.4.0,<3.0.0",
  "torchmetrics>=1.4.1,<2.0.0",
  "hydra-core>=1.3.2,<2.0.0",
  "omegaconf>=2.3.0,<3.0.0",
  "mlflow>=2.15.1,<3.0.0",
  "dvc[s3,gs,gdrive]>=3.55.2,<4.0.0",
  "datasets>=2.20.0,<3.0.0",
  "fire>=0.6.0,<1.0.0",
  "rich>=13.8.0,<14.0.0",
]

[project.scripts]
disaster-tweets = "disaster_tweets_nlp.commands:main"

[tool.poetry]
packages = [{ include = "disaster_tweets_nlp" }]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
ruff = "^0.6.9"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
